Separate data about questions
Now there is data for lots, items and tender questions. Also different
keywords was added to work with different questions. All that was made
to eliminated problem which occured while question was asked only to
tender and wasn't asked to item and lot (there was problem asking them)
and all display checking test cases passed.