Bump package version
This package is in a rolling state. There are no releases and no plans
to make them, so the `.dev` part of the version string is intended to
indicate the "in development" state of the package.
The version part `2.4` is used due to historical reasons (to be greater
than the last used version string).

Disregard the available tags, they exist only to ease the navigation
in commit history.