Rewrite set_to_object()
The new version supports nested objects (i.e. dictionaries
in dictionaries and lists in dictionaries) just like the old one.
Unlike the old version, the new one supports recursion
(lists in dictionaries in lists of dictionaries in dictionaries in
lists of dictionaries ...).
Two-dimensional lists (lists of lists) are not yet supported.